+++ excelerator/__init__.py
@@ -2,7 +2,7 @@
 Library for reading tabular data from excel in a very forgiving manner.
 """
 
-__version__ = "0.4.1"
+__version__ = "0.5"
 
 
 from excelerator.main.table_reader import TableReader