+++ docs/source/change_log.rst
@@ -2,6 +2,12 @@
 Change Log
 ----------
 
+v0.5
+----
+- TableReader:
+    - ``read_from()`` method renamed to ``get_fields()``
+    - new method: ``get_records()``
+
 v0.4
 ----
 - add normalization functions:
@@ -13,4 +19,3 @@ v0.4
 v0.3
 ----
 - add ``header_row_num`` param to ``TableReader``
-